added tab management
now it's possible to move tabs from window to window and to the outside.
yet, the main window needs to be separate from other windows created,
that will enable us to close windows without affecting the application
runtime. [if all the windows need to be equal, will have to develop some
method to verify if all the windows are closed before quitting the app]